Cleanup contract definition of DebugEngineHost
DebugEngineHost had a bunch of different source files, and no XML document comments. This follows the same pattern that CoreFX does where they have one C# file with everything in it for the assembly (since it doesn't take much code to throw NotImplementedException :)). I also added XML doc comments and fixed some spelling mistakes in the comments.

I also enabled FxCop for the contract assembly to make sure 'we' don't mispell anything in an API that will someday be public (and of course by 'we' I am really just thinking of myself :)).
