--- src/DebugEngineHost.Stub/HostLoader.cs
@@ -1,15 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-
-namespace Microsoft.DebugEngineHost
-{
-    public static class HostLoader
-    {
-        public static object VsCoCreateManagedObject(HostConfigurationStore configStore, Guid clsid)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
