--- src/DebugEngineHost.Stub/HostMarshal.cs
@@ -1,26 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using Microsoft.VisualStudio.Debugger.Interop;
-using System;
-
-namespace Microsoft.DebugEngineHost
-{
-    public static class HostMarshal
-    {
-        public static IntPtr RegisterCodeContext(IDebugCodeContext2 codeContext)
-        {
-            throw new NotImplementedException();
-        }
-
-        public static IDebugDocumentPosition2 GetDocumentPositionForIntPtr(IntPtr documentPositionId)
-        {
-            throw new NotImplementedException();
-        }
-
-        public static IDebugEventCallback2 GetThreadSafeEventCallback(IDebugEventCallback2 ad7Callback)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
