--- src/DebugEngineHost.Stub/HostNatvisProject.cs
@@ -1,17 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-
-namespace Microsoft.DebugEngineHost
-{
-    public static class HostNatvisProject
-    {
-        public delegate void NatvisLoader(string path);
-
-        public static void FindNatvisInSolution(NatvisLoader loader)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
